ワーニングが出ないようにした

+++ source/file.hpp
@@ -9,7 +9,7 @@ void read_csv(std::vector<std::vector<double>> &table,
               bool header = false,
               bool index = false,
               bool verbose = true,
-              const std::pair<int, int> size = std::pair<int, int>{0, 0});
+              const std::pair<size_t, size_t> size = std::pair<size_t, size_t>{0, 0});
 void to_csv(const std::vector<double> &table, const std::string path);
 void to_csv(const std::vector<std::vector<double>> &table,
             const std::string path,
